# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "D2C API",
    "description": "D2C owned storefront API service.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/backoffice/site-config/pages/home": {
      "get": {
        "tags": [
          "backoffice-site-config"
        ],
        "summary": "Get Home Config",
        "operationId": "get_home_config_backoffice_site_config_pages_home_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BackofficeHomeConfigResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "backoffice-site-config"
        ],
        "summary": "Patch Home Config",
        "operationId": "patch_home_config_backoffice_site_config_pages_home_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BackofficeHomePagePatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BackofficeHomeConfigResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/backoffice/site-config/pages/home/slots/{slot_code}": {
      "patch": {
        "tags": [
          "backoffice-site-config"
        ],
        "summary": "Patch Home Slot Config",
        "operationId": "patch_home_slot_config_backoffice_site_config_pages_home_slots__slot_code__patch",
        "parameters": [
          {
            "name": "slot_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slot Code"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BackofficeSlotPatchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BackofficeHomeConfigResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/backoffice/site-config/pages/home/slots/{slot_code}/items": {
      "put": {
        "tags": [
          "backoffice-site-config"
        ],
        "summary": "Put Home Slot Items",
        "operationId": "put_home_slot_items_backoffice_site_config_pages_home_slots__slot_code__items_put",
        "parameters": [
          {
            "name": "slot_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slot Code"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BackofficeSlotItemsPutRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BackofficeHomeConfigResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/backoffice/site-config/pages/home/slots/{slot_code}/offer-positions": {
      "put": {
        "tags": [
          "backoffice-site-config"
        ],
        "summary": "Put Home Slot Offer Positions",
        "operationId": "put_home_slot_offer_positions_backoffice_site_config_pages_home_slots__slot_code__offer_positions_put",
        "parameters": [
          {
            "name": "slot_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slot Code"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BackofficeSlotOfferPositionsPutRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BackofficeHomeConfigResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/backoffice/site-config/offers/{offer_code}/resolve": {
      "get": {
        "tags": [
          "backoffice-site-config"
        ],
        "summary": "Resolve Home Offer",
        "operationId": "resolve_home_offer_backoffice_site_config_offers__offer_code__resolve_get",
        "parameters": [
          {
            "name": "offer_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Offer Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BackofficeOfferResolveResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/cart": {
      "get": {
        "tags": [
          "cart"
        ],
        "summary": "Cart Detail",
        "operationId": "cart_detail_cart_get",
        "parameters": [
          {
            "name": "anonymous_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 96,
              "title": "Anonymous Id"
            }
          },
          {
            "name": "session_code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 96,
              "title": "Session Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cart/items": {
      "post": {
        "tags": [
          "cart"
        ],
        "summary": "Cart Items Upsert",
        "operationId": "cart_items_upsert_cart_items_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CartItemUpsertRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cart/clear": {
      "post": {
        "tags": [
          "cart"
        ],
        "summary": "Cart Clear",
        "operationId": "cart_clear_cart_clear_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CartIdentityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BackofficeHomeConfigResponse": {
        "properties": {
          "site": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StorefrontSiteSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "page": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BackofficeHomePage"
              },
              {
                "type": "null"
              }
            ]
          },
          "validation_issues": {
            "items": {
              "$ref": "#/components/schemas/BackofficeValidationIssue"
            },
            "type": "array",
            "title": "Validation Issues"
          }
        },
        "type": "object",
        "title": "BackofficeHomeConfigResponse"
      },
      "BackofficeHomePage": {
        "properties": {
          "page_code": {
            "type": "string",
            "title": "Page Code"
          },
          "page_type": {
            "type": "string",
            "title": "Page Type"
          },
          "route_path": {
            "type": "string",
            "title": "Route Path"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "seo_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seo Title"
          },
          "seo_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seo Description"
          },
          "slots": {
            "items": {
              "$ref": "#/components/schemas/BackofficePageSlot"
            },
            "type": "array",
            "title": "Slots"
          }
        },
        "type": "object",
        "required": [
          "page_code",
          "page_type",
          "route_path",
          "title",
          "status"
        ],
        "title": "BackofficeHomePage"
      },
      "BackofficeHomePagePatchRequest": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 160
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "status": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "active",
                  "disabled"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "seo_title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 180
              },
              {
                "type": "null"
              }
            ],
            "title": "Seo Title"
          },
          "seo_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seo Description"
          }
        },
        "type": "object",
        "title": "BackofficeHomePagePatchRequest"
      },
...
      "BackofficeOfferResolveResponse": {
        "properties": {
          "offer": {
            "$ref": "#/components/schemas/StorefrontHomeOffer"
          },
          "raw": {
            "additionalProperties": true,
            "type": "object",
            "title": "Raw"
          }
        },
        "type": "object",
        "required": [
          "offer"
        ],
        "title": "BackofficeOfferResolveResponse"
      },
...
      "BackofficeSlotItemPut": {
        "properties": {
          "item_code": {
            "type": "string",
            "maxLength": 96,
            "minLength": 1,
            "title": "Item Code"
          },
          "item_type": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Item Type"
          },
          "label": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Label"
          },
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 160
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "subtitle": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 240
              },
              {
                "type": "null"
              }
            ],
            "title": "Subtitle"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "icon": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Icon"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "link_type": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 32
              },
              {
                "type": "null"
              }
            ],
            "title": "Link Type"
          },
          "link_value": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 240
              },
              {
                "type": "null"
              }
            ],
            "title": "Link Value"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          },
          "sort_order": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Sort Order"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "item_code",
          "item_type",
          "label",
          "sort_order"
        ],
        "title": "BackofficeSlotItemPut"
      },
      "BackofficeSlotItemsPutRequest": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/BackofficeSlotItemPut"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "title": "BackofficeSlotItemsPutRequest"
      },
...
      "BackofficeSlotOfferPositionPut": {
        "properties": {
          "position_code": {
            "type": "string",
            "maxLength": 96,
            "minLength": 1,
            "title": "Position Code"
          },
          "offer_code": {
            "type": "string",
            "maxLength": 96,
            "minLength": 1,
            "title": "Offer Code"
          },
          "position_type": {
            "type": "string",
            "maxLength": 64,
            "title": "Position Type",
            "default": "manual"
          },
          "is_featured": {
            "type": "boolean",
            "title": "Is Featured",
            "default": false
          },
          "sort_order": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Sort Order"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          },
          "visible_from": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Visible From"
          },
          "visible_until": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Visible Until"
          }
        },
        "type": "object",
        "required": [
          "position_code",
          "offer_code",
          "sort_order"
        ],
        "title": "BackofficeSlotOfferPositionPut"
      },
      "BackofficeSlotOfferPositionsPutRequest": {
        "properties": {
          "offer_positions": {
            "items": {
              "$ref": "#/components/schemas/BackofficeSlotOfferPositionPut"
            },
            "type": "array",
            "title": "Offer Positions"
          }
        },
        "type": "object",
        "title": "BackofficeSlotOfferPositionsPutRequest"
      },
      "BackofficeSlotPatchRequest": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 160
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "subtitle": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 240
              },
              {
                "type": "null"
              }
            ],
            "title": "Subtitle"
          },
          "content": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "presentation": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Presentation"
          },
          "sort_order": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Sort Order"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          }
        },
        "type": "object",
        "title": "BackofficeSlotPatchRequest"
      },
      "BackofficeValidationIssue": {
        "properties": {
          "level": {
            "type": "string",
            "enum": [
              "error",
              "warning"
            ],
            "title": "Level"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "slot_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slot Code"
          },
          "field_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field Key"
          },
          "offer_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Offer Code"
          }
        },
        "type": "object",
        "required": [
          "level",
          "code",
          "message"
        ],
        "title": "BackofficeValidationIssue"
      },
...
      "CartIdentityRequest": {
        "properties": {
          "anonymous_id": {
            "type": "string",
            "maxLength": 96,
            "minLength": 8,
            "title": "Anonymous Id"
          },
          "session_code": {
            "type": "string",
            "maxLength": 96,
            "minLength": 8,
            "title": "Session Code"
          }
        },
        "type": "object",
        "required": [
          "anonymous_id",
          "session_code"
        ],
        "title": "CartIdentityRequest"
      },
      "CartItemUpsertRequest": {
        "properties": {
          "anonymous_id": {
            "type": "string",
            "maxLength": 96,
            "minLength": 8,
            "title": "Anonymous Id"
          },
          "session_code": {
            "type": "string",
            "maxLength": 96,
            "minLength": 8,
            "title": "Session Code"
          },
          "offer_code": {
            "type": "string",
            "maxLength": 96,
            "minLength": 1,
            "title": "Offer Code"
          },
          "quantity": {
            "type": "integer",
            "maximum": 999.0,
            "minimum": 0.0,
            "title": "Quantity"
          }
        },
        "type": "object",
        "required": [
          "anonymous_id",
          "session_code",
          "offer_code",
          "quantity"
        ],
        "title": "CartItemUpsertRequest"
      },
      "CartLineResponse": {
        "properties": {
          "offer_code": {
            "type": "string",
            "title": "Offer Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity"
          },
          "unit_price_cents": {
            "type": "integer",
            "title": "Unit Price Cents"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "line_subtotal_cents": {
            "type": "integer",
            "title": "Line Subtotal Cents"
          }
        },
        "type": "object",
        "required": [
          "offer_code",
          "name",
          "quantity",
          "unit_price_cents",
          "currency",
          "line_subtotal_cents"
        ],
        "title": "CartLineResponse"
      },
      "CartResponse": {
        "properties": {
          "cart_code": {
            "type": "string",
            "title": "Cart Code"
          },
          "anonymous_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Anonymous Id"
          },
          "session_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Code"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "line_count": {
            "type": "integer",
            "title": "Line Count"
          },
          "item_count": {
            "type": "integer",
            "title": "Item Count"
          },
          "subtotal_cents": {
            "type": "integer",
            "title": "Subtotal Cents"
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/CartLineResponse"
            },
            "type": "array",
            "title": "Lines"
          }
        },
        "type": "object",
        "required": [
          "cart_code",
          "anonymous_id",
          "session_code",
          "currency",
          "line_count",
          "item_count",
          "subtotal_cents",
          "lines"
        ],
        "title": "CartResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "StorefrontSiteSummary": {
        "properties": {
          "site_code": {
            "type": "string",
            "title": "Site Code"
          },
          "site_name": {
            "type": "string",
            "title": "Site Name"
          },
          "brand_name": {
            "type": "string",
            "title": "Brand Name"
          },
          "logo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logo Url"
          },
          "default_currency": {
            "type": "string",
            "title": "Default Currency"
          },
          "theme": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StorefrontTheme"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "site_code",
          "site_name",
          "brand_name",
          "default_currency"
        ],
        "title": "StorefrontSiteSummary"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}